+++ src/app/loadout-builder/generated-sets/CompareDrawer.tsx
@@ -5,9 +5,11 @@ import { t } from 'app/i18next-t';
 import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
 import { DimItem, PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { allItemsSelector, bucketsSelector } from 'app/inventory/selectors';
+import { isPluggableItem } from 'app/inventory/store/sockets';
 import { updateLoadout } from 'app/loadout-drawer/actions';
 import { getItemsFromLoadoutItems } from 'app/loadout-drawer/loadout-item-conversion';
-import { Loadout, LoadoutItem } from 'app/loadout-drawer/loadout-types';
+import { DimLoadoutItem, Loadout, LoadoutItem } from 'app/loadout-drawer/loadout-types';
+import { convertToLoadoutItem } from 'app/loadout-drawer/loadout-utils';
 import { upgradeSpendTierToMaxEnergy } from 'app/loadout/armor-upgrade-utils';
 import Mod from 'app/loadout/loadout-ui/Mod';
 import Sockets from 'app/loadout/loadout-ui/Sockets';
@@ -16,8 +18,10 @@ import { createGetModRenderKey } from 'app/loadout/mod-utils';
 import { useD2Definitions } from 'app/manifest/selectors';
 import { armorStats } from 'app/search/d2-known-values';
 import { useThunkDispatch } from 'app/store/thunk-dispatch';
+import { emptyArray } from 'app/utils/empty';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
+import { BucketHashes } from 'data/d2/generated-enums';
 import produce from 'immer';
 import _ from 'lodash';
 import React, { useMemo, useState } from 'react';
@@ -56,6 +60,7 @@ interface Props {
   loadouts: Loadout[];
   initialLoadoutId?: string;
   lockedMods: PluggableInventoryItemDefinition[];
+  subclass: DimLoadoutItem | undefined;
   classType: DestinyClass;
   statOrder: number[];
   enabledStats: Set<number>;
@@ -87,6 +92,7 @@ export default function CompareDrawer({
   initialLoadoutId,
   set,
   lockedMods,
+  subclass,
   classType,
   statOrder,
   enabledStats,
@@ -111,14 +117,18 @@ export default function CompareDrawer({
   const allItems = useSelector(allItemsSelector);
 
   // This probably isn't needed but I am being cautious as it iterates over the stores.
-  const loadoutItems = useMemo(() => {
+  const { loadoutItems, loadoutSubclass } = useMemo(() => {
     const equippedItems = selectedLoadout?.items.filter((item) => item.equipped);
     const [items] = getItemsFromLoadoutItems(equippedItems, defs, buckets, allItems);
-    return _.sortBy(
+    const loadoutItems = _.sortBy(
       items.filter((item) => LockableBucketHashes.includes(item.bucket.hash)),
       (item) => LockableBucketHashes.indexOf(item.bucket.hash)
     );
-  }, [selectedLoadout, defs, buckets, allItems]);
+    const loadoutSubclass = items.find(
+      (item) => item.bucket.hash === BucketHashes.Subclass && item.classType === classType
+    );
+    return { loadoutItems, loadoutSubclass };
+  }, [selectedLoadout?.items, defs, buckets, allItems, classType]);
 
   const { loSetAssignedMods, itemModAssignments, unassignedMods } = useMemo(() => {
     const { itemModAssignments: loSetAssignedMods } = getCheapestModAssignments(
@@ -161,7 +171,13 @@ export default function CompareDrawer({
     }
   }
 
-  const lockedModStats = getTotalModStatChanges(lockedMods, classType);
+  const subclassPlugs = subclass?.socketOverrides
+    ? Object.values(subclass.socketOverrides)
+        .map((hash) => defs.InventoryItem.get(hash))
+        .filter(isPluggableItem)
+    : emptyArray<PluggableInventoryItemDefinition>();
+
+  const lockedModStats = getTotalModStatChanges(lockedMods, subclassPlugs, classType);
 
   for (const statHash of armorStats) {
     loadoutStats[statHash] += lockedModStats[statHash];
@@ -179,20 +195,22 @@ export default function CompareDrawer({
 
     const loadoutToSave = produce(selectedLoadout, (draftLoadout) => {
       if (draftLoadout) {
-        const newItems: LoadoutItem[] = setItems.map(({ id, hash }) => ({
-          id,
-          hash,
-          amount: 1,
-          equipped: true,
-        }));
+        const newItems: LoadoutItem[] = setItems.map((item) => convertToLoadoutItem(item, true));
+
+        if (subclass) {
+          newItems.push(convertToLoadoutItem(subclass, true));
+        }
+
         for (const item of draftLoadout.items) {
           const dimItem = loadoutItems.find((i) => i.id === item.id);
           const hasBeenReplaced =
-            dimItem && setItems.some((i) => i.bucket.hash === dimItem.bucket.hash);
+            (dimItem && setItems.some((i) => i.bucket.hash === dimItem.bucket.hash)) ||
+            (subclass && item.id === loadoutSubclass?.id);
           if (!hasBeenReplaced) {
             newItems.push(item);
           }
         }
+
         draftLoadout.items = newItems;
         draftLoadout.parameters = params;
         draftLoadout.notes = notes || draftLoadout.notes;